added "steps" display
this should much better represent the output from s0 electricity meters
and water meters, where we get one impuls per energy or volume unit and
(more important) the time interval depends on the throughput.
this does not make much sense for temperature, though. unfortunately,
we have currently no means to select different chart types depending
on the entity type.
also, I had to change the backend to output the starting time of the
interval (flot needs that) instead of the ending time, which changes
the "lines" chart. but since this chart is wrong for electricity and
water anyway, it doesn't realy matter, though.